test(editor): give the SQL fallback real provider coverage (N2)
Addresses the remaining review gap: the N2/D7 block asserted the resolver
directly and only that the component mounts, so CodeQueryEditor could keep its
local seven-entry fallback and still pass. Replaced with provider invocations
using an omitted suggestions prop, asserting sum/avg/count/max/min/histogram/
approx_topk and the array family come back, Function-kinded and snippet-enabled.
Fails today with 'fallback missing sum'.

Moved the whole provider suite into CodeQueryEditor.completion.spec.ts. In the
host file the pre-existing load flakiness (Ctrl+Enter, setValue) starved this
suite's beforeAll and vitest SKIPPED all of it — silently zero coverage.
Isolated, it is stable: 2 passed / 10 failed across three consecutive runs.
